ENH: Split vtk<->igtl conversion from vtkIGTLToMRMLImage into a separate codec class igtlCodecImage.
This is the first step in an attempt to split out all vtk<->igtl code from the OpenIGTLinkIF repo and
eventually move it to the OpenIGTLink repo. See
http://www.na-mic.org/Wiki/index.php/2016_Winter_Project_Week/Projects/TrackedUltrasoundStandardization
for background.

Note: this commit compiles on mac but it not yet tested. Beware!
