+++ MRML/CMakeLists.txt
@@ -47,6 +47,7 @@ set(${KIT}_TARGET_LIBRARIES
   SlicerBaseLogic
   vtkSlicerAnnotationsModuleMRML
   vtkSlicerMarkupsModuleMRML
+  igtlSupportModuleCodec
   )
 
 #-----------------------------------------------------------------------------
