+++ MRML/vtkIGTLToMRMLImage.cxx
@@ -16,6 +16,9 @@
 #include "vtkIGTLToMRMLImage.h"
 #include "vtkMRMLIGTLQueryNode.h"
 
+// igtl support includes
+#include <igtlCodecImage.h>
+
 // OpenIGTLink includes
 #include <igtl_util.h>
 #include <igtlImageMessage.h>
@@ -206,46 +209,6 @@ vtkIntArray* vtkIGTLToMRMLImage::GetNodeEvents()
   return events;
 }
 
-
-//---------------------------------------------------------------------------
-// Stream copy + byte swap
-//---------------------------------------------------------------------------
-int swapCopy16(igtlUint16 * dst, igtlUint16 * src, int n)
-{
-  igtlUint16 * esrc = src + n;
-  while (src < esrc)
-    {
-    *dst = BYTE_SWAP_INT16(*src);
-    dst ++;
-    src ++;
-    }
-  return 1;
-}
-
-int swapCopy32(igtlUint32 * dst, igtlUint32 * src, int n)
-{
-  igtlUint32 * esrc = src + n;
-  while (src < esrc)
-    {
-    *dst = BYTE_SWAP_INT32(*src);
-    dst ++;
-    src ++;
-    }
-  return 1;
-}
-
-int swapCopy64(igtlUint64 * dst, igtlUint64 * src, int n)
-{
-  igtlUint64 * esrc = src + n;
-  while (src < esrc)
-    {
-    *dst = BYTE_SWAP_INT64(*src);
-    dst ++;
-    src ++;
-    }
-  return 1;
-}
-
 //---------------------------------------------------------------------------
 int vtkIGTLToMRMLImage::IGTLToMRML(igtl::MessageBase::Pointer buffer, vtkMRMLNode* node)
 {
@@ -256,303 +219,67 @@ int vtkIGTLToMRMLImage::IGTLToMRML(igtl::MessageBase::Pointer buffer, vtkMRMLNod
     return 0;
     }
 
-  // Create a message buffer to receive image data
-  igtl::ImageMessage::Pointer imgMsg;
-  imgMsg = igtl::ImageMessage::New();
-  imgMsg->Copy(buffer); // !! TODO: copy makes performance issue.
+  if (!Codec.Get())
+    Codec = vtkSmartPointer<igtlCodecImage>::New();
 
-  // Deserialize the transform data
-  // If CheckCRC==0, CRC check is skipped.
-  int c = imgMsg->Unpack(this->CheckCRC);
+  igtlCodecImage::MessageContent content;
+  content.image = volumeNode->GetImageData(); // reuse existing imagedata if structure (dims, type, components) is identical
 
-  if ((c & igtl::MessageHeader::UNPACK_BODY) == 0) // if CRC check fails
+  if (Codec->IGTLToVTK(buffer, &content, this->CheckCRC) == 0)
     {
-    // TODO: error handling
     return 0;
     }
 
-  // Retrieve the image data
-  int   size[3];          // image dimension
-  float spacing[3];       // spacing (mm/pixel)
-  int   svsize[3];        // sub-volume size
-  int   svoffset[3];      // sub-volume offset
-  int   scalarType;       // VTK scalar type
-  int   numComponents;    // number of scalar components
-  int   endian;
-  igtl::Matrix4x4 matrix; // Image origin and orientation matrix
+  vtkSmartPointer<vtkImageData> nodeImageData = volumeNode->GetImageData();
 
-  scalarType = IGTLToVTKScalarType( imgMsg->GetScalarType() );
-  endian = imgMsg->GetEndian();
-  imgMsg->GetDimensions(size);
-  imgMsg->GetSpacing(spacing);
-  numComponents = imgMsg->GetNumComponents();
-  imgMsg->GetSubVolume(svsize, svoffset);
-  imgMsg->GetMatrix(matrix);
+  int numComponentsInNode = 0;
+  int numComponents = 0;
 
-  // check if the IGTL data fits to the current MRML node  
-  int sizeInNode[3]={0,0,0};
-  int scalarTypeInNode=VTK_VOID;
-  int numComponentsInNode=0;
-  // Get vtk image from MRML node
-  vtkSmartPointer<vtkImageData> imageData = volumeNode->GetImageData();
-  if (imageData.GetPointer()!=NULL)
+  if (nodeImageData)
+    numComponentsInNode = nodeImageData->GetNumberOfScalarComponents();
+  numComponents = content.image->GetNumberOfScalarComponents();
+
+  if (numComponentsInNode != numComponents)
     {
-    imageData->GetDimensions(sizeInNode);
-    scalarTypeInNode = imageData->GetScalarType();
-    numComponentsInNode = imageData->GetNumberOfScalarComponents();
-    }
+    // number of components changed, so we need to remove the incompatible
+    // dispay nodes and create a default display node if no compatible display
+    // node remains
 
-  if (imageData.GetPointer()==NULL
-      || sizeInNode[0] != size[0] || sizeInNode[1] != size[1] || sizeInNode[2] != size[2]
-      || scalarType != scalarTypeInNode
-      || numComponentsInNode != numComponents)
-    {
-    if (numComponentsInNode != numComponents)
+    bool scalarDisplayNodeRequired = (numComponents==1);
+    bool mayNeedToRemoveDisplayNodes=false;
+    do
       {
-      // number of components changed, so we need to remove the incompatible
-      // dispay nodes and create a default display node if no compatible display
-      // node remains
-      
-      bool scalarDisplayNodeRequired = (numComponents==1);
-      bool mayNeedToRemoveDisplayNodes=false;
-      do
+      mayNeedToRemoveDisplayNodes=false;
+      for (int i=0; i<volumeNode->GetNumberOfDisplayNodes(); i++)
         {
-        mayNeedToRemoveDisplayNodes=false;
-        for (int i=0; i<volumeNode->GetNumberOfDisplayNodes(); i++)
+        vtkMRMLVolumeDisplayNode* currentDisplayNode = vtkMRMLVolumeDisplayNode::SafeDownCast(volumeNode->GetNthDisplayNode(i));
+        bool currentDisplayNodeIsScalar = (vtkMRMLVectorVolumeDisplayNode::SafeDownCast(currentDisplayNode)==NULL);
+        if (scalarDisplayNodeRequired!=currentDisplayNodeIsScalar)
           {
-          vtkMRMLVolumeDisplayNode* currentDisplayNode = vtkMRMLVolumeDisplayNode::SafeDownCast(volumeNode->GetNthDisplayNode(i));
-          bool currentDisplayNodeIsScalar = (vtkMRMLVectorVolumeDisplayNode::SafeDownCast(currentDisplayNode)==NULL);
-          if (scalarDisplayNodeRequired!=currentDisplayNodeIsScalar)
-            {
-            // incompatible display node, remove it
-            //volumeNode->RemoveNthDisplayNodeID(i);
-            volumeNode->GetScene()->RemoveNode(currentDisplayNode);
-            mayNeedToRemoveDisplayNodes=true;
-            }
-          }
-        }
-      while (mayNeedToRemoveDisplayNodes);
-
-      if (volumeNode->GetNumberOfDisplayNodes()==0)
-        {
-        // the new default display node may be incompatible with the current image data,
-        // so clear it to make sure no display is attempted with incompatible image data
-        volumeNode->SetAndObserveImageData(NULL);
-        SetDefaultDisplayNode(volumeNode, numComponents);
-        }
-      }
-    
-    imageData = vtkSmartPointer<vtkImageData>::New();
-    imageData->SetDimensions(size[0], size[1], size[2]);
-    imageData->SetExtent(0, size[0]-1, 0, size[1]-1, 0, size[2]-1);
-    imageData->SetOrigin(0.0, 0.0, 0.0);
-    imageData->SetSpacing(1.0, 1.0, 1.0);
-#if (VTK_MAJOR_VERSION <= 5)
-    imageData->SetNumberOfScalarComponents(numComponents);
-    imageData->SetScalarType(scalarType);
-    imageData->AllocateScalars();
-#else
-    imageData->AllocateScalars(scalarType, numComponents);
-#endif
-    }
-  
-  float tx = matrix[0][0];
-  float ty = matrix[1][0];
-  float tz = matrix[2][0];
-  float sx = matrix[0][1];
-  float sy = matrix[1][1];
-  float sz = matrix[2][1];
-  float nx = matrix[0][2];
-  float ny = matrix[1][2];
-  float nz = matrix[2][2];
-  float px = matrix[0][3];
-  float py = matrix[1][3];
-  float pz = matrix[2][3];
-
-  // Check scalar size
-  int scalarSize = imgMsg->GetScalarSize();
-  
-  int fByteSwap = 0;
-  // Check if bytes-swap is required
-  if (scalarSize > 1 && 
-      ((igtl_is_little_endian() && endian == igtl::ImageMessage::ENDIAN_BIG) ||
-       (!igtl_is_little_endian() && endian == igtl::ImageMessage::ENDIAN_LITTLE)))
-    {
-    // Needs byte swap
-    fByteSwap = 1;
-    }
-
-  if (imgMsg->GetImageSize() == imgMsg->GetSubVolumeImageSize())
-    {
-    // In case that volume size == sub-volume size,
-    // image is read directly to the memory area of vtkImageData
-    // for better performance. 
-    if (fByteSwap)
-      {
-      switch (scalarSize)
-        {
-        case 2:
-          swapCopy16((igtlUint16 *)imageData->GetScalarPointer(),
-                     (igtlUint16 *)imgMsg->GetScalarPointer(),
-                     imgMsg->GetSubVolumeImageSize() / 2);
-          break;
-        case 4:
-          swapCopy32((igtlUint32 *)imageData->GetScalarPointer(),
-                     (igtlUint32 *)imgMsg->GetScalarPointer(),
-                     imgMsg->GetSubVolumeImageSize() / 4);
-          break;
-        case 8:
-          swapCopy64((igtlUint64 *)imageData->GetScalarPointer(),
-                     (igtlUint64 *)imgMsg->GetScalarPointer(),
-                     imgMsg->GetSubVolumeImageSize() / 8);
-          break;
-        default:
-          break;
-        }
-      }
-    else
-      {
-      memcpy(imageData->GetScalarPointer(),
-             imgMsg->GetScalarPointer(), imgMsg->GetSubVolumeImageSize());
-      }
-    }
-  else
-    {
-    // In case of volume size != sub-volume size,
-    // image is loaded into ImageReadBuffer, then copied to
-    // the memory area of vtkImageData.
-    char* imgPtr = (char*) imageData->GetScalarPointer();
-    char* bufPtr = (char*) imgMsg->GetScalarPointer();
-    int sizei = size[0];
-    int sizej = size[1];
-    //int sizek = size[2];
-    int subsizei = svsize[0];
-    
-    int bg_i = svoffset[0];
-    //int ed_i = bg_i + svsize[0];
-    int bg_j = svoffset[1];
-    int ed_j = bg_j + svsize[1];
-    int bg_k = svoffset[2];
-    int ed_k = bg_k + svsize[2];
-
-    if (fByteSwap)
-      {
-      switch (scalarSize)
-        {
-        case 2:
-          for (int k = bg_k; k < ed_k; k ++)
-            {
-            for (int j = bg_j; j < ed_j; j ++)
-              {
-              swapCopy16((igtlUint16 *)&imgPtr[(sizei*sizej*k + sizei*j + bg_i)*scalarSize],
-                         (igtlUint16 *)bufPtr,
-                         subsizei);
-              bufPtr += subsizei*scalarSize;
-              }
-            }
-          break;
-        case 4:
-          for (int k = bg_k; k < ed_k; k ++)
-            {
-            for (int j = bg_j; j < ed_j; j ++)
-              {
-              swapCopy32((igtlUint32 *)&imgPtr[(sizei*sizej*k + sizei*j + bg_i)*scalarSize],
-                         (igtlUint32 *)bufPtr,
-                         subsizei);
-              bufPtr += subsizei*scalarSize;
-              }
-            }
-          break;
-        case 8:
-          for (int k = bg_k; k < ed_k; k ++)
-            {
-            for (int j = bg_j; j < ed_j; j ++)
-              {
-              swapCopy64((igtlUint64 *)&imgPtr[(sizei*sizej*k + sizei*j + bg_i)*scalarSize],
-                         (igtlUint64 *)bufPtr,
-                         subsizei);
-              bufPtr += subsizei*scalarSize;
-              }
-            }
-          break;
-        default:
-          break;
-        }
-      }
-    else
-      {
-      for (int k = bg_k; k < ed_k; k ++)
-        {
-        for (int j = bg_j; j < ed_j; j ++)
-          {
-          memcpy(&imgPtr[(sizei*sizej*k + sizei*j + bg_i)*scalarSize],
-                 bufPtr, subsizei*scalarSize);
-          bufPtr += subsizei*scalarSize;
+          // incompatible display node, remove it
+          //volumeNode->RemoveNthDisplayNodeID(i);
+          volumeNode->GetScene()->RemoveNode(currentDisplayNode);
+          mayNeedToRemoveDisplayNodes=true;
           }
         }
       }
+    while (mayNeedToRemoveDisplayNodes);
 
+    if (volumeNode->GetNumberOfDisplayNodes()==0)
+      {
+      // the new default display node may be incompatible with the current image data,
+      // so clear it to make sure no display is attempted with incompatible image data
+      volumeNode->SetAndObserveImageData(NULL);
+      SetDefaultDisplayNode(volumeNode, numComponents);
+      }
     }
-  
-  // normalize
-  float psi = sqrt(tx*tx + ty*ty + tz*tz);
-  float psj = sqrt(sx*sx + sy*sy + sz*sz);
-  float psk = sqrt(nx*nx + ny*ny + nz*nz);
-  float ntx = tx / psi;
-  float nty = ty / psi;
-  float ntz = tz / psi;
-  float nsx = sx / psj;
-  float nsy = sy / psj;
-  float nsz = sz / psj;
-  float nnx = nx / psk;
-  float nny = ny / psk;
-  float nnz = nz / psk;
 
-  // Shift the center
-  // NOTE: The center of the image should be shifted due to different
-  // definitions of image origin between VTK (Slicer) and OpenIGTLink;
-  // OpenIGTLink image has its origin at the center, while VTK image
-  // has one at the corner.
-  float hfovi = spacing[0] * psi * (size[0]-1) / 2.0;
-  float hfovj = spacing[1] * psj * (size[1]-1) / 2.0;
-  float hfovk = spacing[2] * psk * (size[2]-1) / 2.0;
-  //float hfovk = 0;
-
-  float cx = ntx * hfovi + nsx * hfovj + nnx * hfovk;
-  float cy = nty * hfovi + nsy * hfovj + nny * hfovk;
-  float cz = ntz * hfovi + nsz * hfovj + nnz * hfovk;
-  px = px - cx;
-  py = py - cy;
-  pz = pz - cz;
-
-  // set volume orientation
-  vtkMatrix4x4* rtimgTransform = vtkMatrix4x4::New();
-  rtimgTransform->Identity();
-  rtimgTransform->Element[0][0] = ntx*spacing[0];
-  rtimgTransform->Element[1][0] = nty*spacing[0];
-  rtimgTransform->Element[2][0] = ntz*spacing[0];
-  rtimgTransform->Element[0][1] = nsx*spacing[1];
-  rtimgTransform->Element[1][1] = nsy*spacing[1];
-  rtimgTransform->Element[2][1] = nsz*spacing[1];
-  rtimgTransform->Element[0][2] = nnx*spacing[2];
-  rtimgTransform->Element[1][2] = nny*spacing[2];
-  rtimgTransform->Element[2][2] = nnz*spacing[2];
-  rtimgTransform->Element[0][3] = px;
-  rtimgTransform->Element[1][3] = py;
-  rtimgTransform->Element[2][3] = pz;
-
-  //rtimgTransform->Invert();
-  //volumeNode->SetRASToIJKMatrix(rtimgTransform);
-  volumeNode->SetIJKToRASMatrix(rtimgTransform);
-  rtimgTransform->Delete();
-
-  volumeNode->SetAndObserveImageData(imageData);
-
-  imageData->Modified();
+  // insert igtl transform and image into MRML:
+  volumeNode->SetIJKToRASMatrix(content.transform);
+  volumeNode->SetAndObserveImageData(content.image);
   volumeNode->Modified();
 
   return 1;
-
 }
 
 //---------------------------------------------------------------------------
@@ -567,106 +294,23 @@ int vtkIGTLToMRMLImage::MRMLToIGTL(unsigned long event, vtkMRMLNode* mrmlNode, i
   if (event == vtkMRMLVolumeNode::ImageDataModifiedEvent && strcmp(mrmlNode->GetNodeTagName(), "Volume") == 0)
     {
     vtkMRMLScalarVolumeNode* volumeNode =
-      vtkMRMLScalarVolumeNode::SafeDownCast(mrmlNode);
+        vtkMRMLScalarVolumeNode::SafeDownCast(mrmlNode);
 
     if (!volumeNode)
       {
       return 0;
       }
 
-    vtkImageData* imageData = volumeNode->GetImageData();
-    int   isize[3];          // image dimension
-    //int   svsize[3];        // sub-volume size
-    int   scalarType;       // scalar type
-    //double *origin;
-    double *spacing;       // spacing (mm/pixel)
-    int   ncomp;
-    int   svoffset[] = {0, 0, 0};           // sub-volume offset
-    int   endian;
+    if (!Codec.Get())
+      Codec = vtkSmartPointer<igtlCodecImage>::New();
 
-    scalarType = imageData->GetScalarType();
-    ncomp = imageData->GetNumberOfScalarComponents();
-    imageData->GetDimensions(isize);
-    //imageData->GetExtent(0, isize[0]-1, 0, isize[1]-1, 0, isize[2]-1);
-    //origin = imageData->GetOrigin();
-    spacing = imageData->GetSpacing();
+    igtlCodecImage::MessageContent content;
+    content.image = volumeNode->GetImageData();
 
-    // Check endianness of the machine
-    endian = igtl::ImageMessage::ENDIAN_BIG;
-    if (igtl_is_little_endian())
+    if (Codec->VTKToIGTL(content, &this->OutImageMessage) == 0)
       {
-      endian = igtl::ImageMessage::ENDIAN_LITTLE;
+      return 0;
       }
-
-    if (this->OutImageMessage.IsNull())
-      {
-      this->OutImageMessage = igtl::ImageMessage::New();
-      }
-    this->OutImageMessage->SetDimensions(isize);
-    this->OutImageMessage->SetSpacing((float)spacing[0], (float)spacing[1], (float)spacing[2]);
-    this->OutImageMessage->SetScalarType(scalarType);
-    this->OutImageMessage->SetEndian(endian);
-    this->OutImageMessage->SetDeviceName(volumeNode->GetName());
-    this->OutImageMessage->SetSubVolume(isize, svoffset);
-    this->OutImageMessage->SetNumComponents(ncomp);
-    this->OutImageMessage->AllocateScalars();
-
-    memcpy(this->OutImageMessage->GetScalarPointer(),
-           imageData->GetScalarPointer(),
-           this->OutImageMessage->GetImageSize());
-
-    // Transform
-    vtkMatrix4x4* rtimgTransform = vtkMatrix4x4::New();
-    volumeNode->GetIJKToRASMatrix(rtimgTransform);
-    float ntx = rtimgTransform->Element[0][0] / (float)spacing[0];
-    float nty = rtimgTransform->Element[1][0] / (float)spacing[0];
-    float ntz = rtimgTransform->Element[2][0] / (float)spacing[0];
-    float nsx = rtimgTransform->Element[0][1] / (float)spacing[1];
-    float nsy = rtimgTransform->Element[1][1] / (float)spacing[1];
-    float nsz = rtimgTransform->Element[2][1] / (float)spacing[1];
-    float nnx = rtimgTransform->Element[0][2] / (float)spacing[2];
-    float nny = rtimgTransform->Element[1][2] / (float)spacing[2];
-    float nnz = rtimgTransform->Element[2][2] / (float)spacing[2];
-    float px  = rtimgTransform->Element[0][3];
-    float py  = rtimgTransform->Element[1][3];
-    float pz  = rtimgTransform->Element[2][3];
-
-    rtimgTransform->Delete();
-
-    // Shift the center
-    // NOTE: The center of the image should be shifted due to different
-    // definitions of image origin between VTK (Slicer) and OpenIGTLink;
-    // OpenIGTLink image has its origin at the center, while VTK image
-    // has one at the corner.
-
-    float hfovi = (float)spacing[0] * (float)(isize[0]-1) / 2.0;
-    float hfovj = (float)spacing[1] * (float)(isize[1]-1) / 2.0;
-    float hfovk = (float)spacing[2] * (float)(isize[2]-1) / 2.0;
-
-    float cx = ntx * hfovi + nsx * hfovj + nnx * hfovk;
-    float cy = nty * hfovi + nsy * hfovj + nny * hfovk;
-    float cz = ntz * hfovi + nsz * hfovj + nnz * hfovk;
-
-    px = px + cx;
-    py = py + cy;
-    pz = pz + cz;
-
-    igtl::Matrix4x4 matrix; // Image origin and orientation matrix
-    matrix[0][0] = ntx;
-    matrix[1][0] = nty;
-    matrix[2][0] = ntz;
-    matrix[0][1] = nsx;
-    matrix[1][1] = nsy;
-    matrix[2][1] = nsz;
-    matrix[0][2] = nnx;
-    matrix[1][2] = nny;
-    matrix[2][2] = nnz;
-    matrix[0][3] = px;
-    matrix[1][3] = py;
-    matrix[2][3] = pz;
-
-    this->OutImageMessage->SetMatrix(matrix);
-    this->OutImageMessage->Pack();
 
     *size = this->OutImageMessage->GetPackSize();
     *igtlMsg = (void*)this->OutImageMessage->GetPackPointer();
@@ -693,15 +337,15 @@ int vtkIGTLToMRMLImage::MRMLToIGTL(unsigned long event, vtkMRMLNode* mrmlNode, i
       /*
       else if (qnode->GetQueryType() == vtkMRMLIGTLQueryNode::TYPE_START)
         {
-        *size = 0;
+       *size = 0;
         return 0;
         }
       else if (qnode->GetQueryType() == vtkMRMLIGTLQueryNode::TYPE_STOP)
         {
-        *size = 0;
+       *size = 0;
         return 0;
         }
-      */
+       */
       return 0;
       }
     }
@@ -716,54 +360,35 @@ int vtkIGTLToMRMLImage::MRMLToIGTL(unsigned long event, vtkMRMLNode* mrmlNode, i
 //---------------------------------------------------------------------------
 void vtkIGTLToMRMLImage::CenterImage(vtkMRMLVolumeNode *volumeNode)
 {
-    if ( volumeNode )
-      {
-      vtkImageData *image = volumeNode->GetImageData();
-      if (image)
-        {
-        vtkMatrix4x4 *ijkToRAS = vtkMatrix4x4::New();
-        volumeNode->GetIJKToRASMatrix(ijkToRAS);
-
-        double dimsH[4];
-        double rasCorner[4];
-        int *dims = image->GetDimensions();
-        dimsH[0] = dims[0] - 1;
-        dimsH[1] = dims[1] - 1;
-        dimsH[2] = dims[2] - 1;
-        dimsH[3] = 0.;
-        ijkToRAS->MultiplyPoint(dimsH, rasCorner);
-
-        double origin[3];
-        int i;
-        for (i = 0; i < 3; i++)
-          {
-          origin[i] = -0.5 * rasCorner[i];
-          }
-        volumeNode->SetDisableModifiedEvent(1);
-        volumeNode->SetOrigin(origin);
-        volumeNode->SetDisableModifiedEvent(0);
-        volumeNode->InvokePendingModifiedEvent();
-
-        ijkToRAS->Delete();
-        }
-      }
-}
-
-//---------------------------------------------------------------------------
-int vtkIGTLToMRMLImage::IGTLToVTKScalarType(int igtlType)
-{
-  switch (igtlType)
+  if ( volumeNode )
     {
-    case igtl::ImageMessage::TYPE_INT8: return VTK_CHAR;
-    case igtl::ImageMessage::TYPE_UINT8: return VTK_UNSIGNED_CHAR;
-    case igtl::ImageMessage::TYPE_INT16: return VTK_SHORT;
-    case igtl::ImageMessage::TYPE_UINT16: return VTK_UNSIGNED_SHORT;
-    case igtl::ImageMessage::TYPE_INT32: return VTK_UNSIGNED_LONG;
-    case igtl::ImageMessage::TYPE_UINT32: return VTK_UNSIGNED_LONG;
-    case igtl::ImageMessage::TYPE_FLOAT32: return VTK_FLOAT;
-    case igtl::ImageMessage::TYPE_FLOAT64: return VTK_DOUBLE;
-    default:
-      vtkErrorMacro ("Invalid IGTL scalar Type: "<<igtlType);
-      return VTK_VOID;
+    vtkImageData *image = volumeNode->GetImageData();
+    if (image)
+      {
+      vtkMatrix4x4 *ijkToRAS = vtkMatrix4x4::New();
+      volumeNode->GetIJKToRASMatrix(ijkToRAS);
+
+      double dimsH[4];
+      double rasCorner[4];
+      int *dims = image->GetDimensions();
+      dimsH[0] = dims[0] - 1;
+      dimsH[1] = dims[1] - 1;
+      dimsH[2] = dims[2] - 1;
+      dimsH[3] = 0.;
+      ijkToRAS->MultiplyPoint(dimsH, rasCorner);
+
+      double origin[3];
+      int i;
+      for (i = 0; i < 3; i++)
+        {
+        origin[i] = -0.5 * rasCorner[i];
+        }
+      volumeNode->SetDisableModifiedEvent(1);
+      volumeNode->SetOrigin(origin);
+      volumeNode->SetDisableModifiedEvent(0);
+      volumeNode->InvokePendingModifiedEvent();
+
+      ijkToRAS->Delete();
+      }
     }
 }
